requirements.txt

torch==2.1.0
torchvision==0.16.0
diffusers==0.24.0
transformers==4.36.0
open_clip_torch==2.20.0
flask==2.3.3
pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.0.76
peft==0.5.0
accelerate==0.23.0
tqdm==4.66.1
safetensors==0.4.0
einops==0.6.1